add module class

+++ Modules/MasterData/Http/Controllers/SiswaController.php
@@ -2,13 +2,17 @@
 
 namespace Modules\MasterData\Http\Controllers;
 
+use App\Exports\SiswaExport;
 use App\Imports\SiswaImport;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use App\Exports\SiswaCategoryExport;
 use Maatwebsite\Excel\Facades\Excel;
+use Modules\MasterData\Entities\Kelas;
 use Modules\MasterData\Entities\Siswa;
 use RealRashid\SweetAlert\Facades\Alert;
 use Illuminate\Support\Facades\Validator;
+use Modules\MasterData\Entities\Category;
 use Modules\MasterData\Entities\TahunMasuk;
 use Illuminate\Contracts\Support\Renderable;
 
@@ -18,17 +22,26 @@ class SiswaController extends Controller
    * Display a listing of the resource.
    * @return Renderable
    */
-  public function index()
+  public function index(Request $request)
   {
     $title = 'Data Siswa';
-    $data = Siswa::with('tahunMasuk')->get();
+    $search = $request->input('search');
 
-    $alert = 'Delete Data!';
-    $text = "Are you sure you want to delete?";
-    confirmDelete($alert, $text);
+    if ($search) {
+      $data = Siswa::with(['tahunMasuk', 'category'])
+        ->where('category_id', 'like', '%' . $search . '%')
+        ->latest()
+        ->get();
+    } else {
+      $data = Siswa::with(['tahunMasuk', 'category'])->get();
+    }
 
-    return view('masterdata::siswa.index', ['data' => $data, 'title' => $title]);
+    return view('masterdata::siswa.index', [
+      'data' => $data,
+      'title' => $title
+    ]);
   }
+
 
   /**
    * Show the form for creating a new resource.
@@ -38,10 +51,16 @@ class SiswaController extends Controller
   {
     $title = "Siswa Baru";
     $data = TahunMasuk::latest()->get();
+    $data_category = Category::with('kelas')->get();
+    $data_kelas = Kelas::latest()->get();
 
-    return view('masterdata::siswa.create', ['data' => $data, 'title' => $title]);
+    return view('masterdata::siswa.create', [
+      'data' => $data,
+      'title' => $title,
+      'data_category' => $data_category,
+      'data_kelas' => $data_kelas
+    ]);
   }
-
   /**
    * Store a newly created resource in storage.
    * @param Request $request
@@ -49,34 +68,35 @@ class SiswaController extends Controller
    */
   public function store(Request $request)
   {
-    // Validasi input
     $validator = Validator::make($request->all(), [
-      // 'name' => 'required|string|max:255',
-      // 'nisn' => 'required|numeric|digits:8',
       'tahun_masuk_id' => 'required|exists:tahun_masuks,id',
     ], [
       'name.required' => 'Nama wajib diisi.',
-      'nisn.required' => 'NISN wajib diisi.',
-      'nisn.numeric' => 'NISN harus berupa angka.',
-      'nisn.digits' => 'NISN harus terdiri dari 8 digit.',
       'tahun_masuk_id.required' => 'Tahun Masuk wajib dipilih.',
       'tahun_masuk_id.exists' => 'Tahun Masuk yang dipilih tidak valid.',
     ]);
 
-    // Jika validasi gagal, kembalikan ke halaman sebelumnya dengan error
     if ($validator->fails()) {
       return redirect()->back()
         ->withErrors($validator)
         ->withInput();
     }
 
-    // Simpan data ke database
-    $newData = Siswa::create($request->all());
+    list($category_id, $kelas_id) = explode('_', $request->input('category_id_kelas_id'));
 
-    // Tampilkan pesan sukses
+    $newData = Siswa::create([
+      'name' => $request->input('name'),
+      'nisn' => $request->input('nisn'),
+      'no_urut' => $request->input('no_urut'),
+      'tahun_masuk_id' => $request->input('tahun_masuk_id'),
+      'category_id' => $category_id,
+      'kelas_id' => $kelas_id,
+    ]);
+
     Alert::success('Success', 'Data berhasil disimpan');
     return redirect()->route('siswa.index');
   }
+
 
 
   /**
@@ -96,11 +116,18 @@ class SiswaController extends Controller
    */
   public function edit($id)
   {
-    $title = "Update SIswa";
+    $title = "Update Siswa";
     $data = Siswa::findOrFail($id);
     $tahun_masuk = TahunMasuk::latest()->get();
-    return view('masterdata::siswa.edit', ['tahun_masuk' => $tahun_masuk, 'data' => $data, 'title' => $title]);
+    $data_category = Category::with('kelas')->get(); // Pastikan juga mengirim data category dan kelas
+    return view('masterdata::siswa.edit', [
+      'tahun_masuk' => $tahun_masuk,
+      'data' => $data,
+      'title' => $title,
+      'data_category' => $data_category
+    ]);
   }
+
 
   /**
    * Update the specified resource in storage.
@@ -110,12 +137,37 @@ class SiswaController extends Controller
    */
   public function update(Request $request, $id)
   {
-    $data = $request->all();
+    $validator = Validator::make($request->all(), [
+      'tahun_masuk_id' => 'required|exists:tahun_masuks,id',
+    ], [
+      'name.required' => 'Nama wajib diisi.',
+      'tahun_masuk_id.required' => 'Tahun Masuk wajib dipilih.',
+      'tahun_masuk_id.exists' => 'Tahun Masuk yang dipilih tidak valid.',
+    ]);
+
+    if ($validator->fails()) {
+      return redirect()->back()
+        ->withErrors($validator)
+        ->withInput();
+    }
+
     $updateData = Siswa::findOrFail($id);
-    $updateData->update($data);
+
+    list($category_id, $kelas_id) = explode('_', $request->input('category_id_kelas_id'));
+
+    $updateData->update([
+      'name' => $request->input('name'),
+      'nisn' => $request->input('nisn'),
+      'no_urut' => $request->input('no_urut'),
+      'tahun_masuk_id' => $request->input('tahun_masuk_id'),
+      'category_id' => $category_id,
+      'kelas_id' => $kelas_id,
+    ]);
+
     Alert::success('Success', 'Data berhasil diupdate');
     return redirect()->route('siswa.index');
   }
+
 
   /**
    * Remove the specified resource from storage.
@@ -125,10 +177,10 @@ class SiswaController extends Controller
   public function destroy($id)
   {
     $data = Siswa::findOrFail($id);
-    // if ($data->tagihans()->count() > 0) {
-    //   Alert::error('Oops....', 'Data tidak dapat dihapus karena memiliki tagihan');
-    //   return redirect()->route('siswa.index');
-    // }
+    if ($data->tagihans()->count() > 0) {
+      Alert::error('Oops....', 'Data tidak dapat dihapus karena memiliki tagihan');
+      return redirect()->route('siswa.index');
+    }
     $data->delete();
     Alert::success('Success', 'Data berhasil dihapus');
     return redirect()->route('siswa.index');
@@ -137,32 +189,54 @@ class SiswaController extends Controller
   public function createImport()
   {
     $title = 'Data Siswa';
-    return view('masterdata::siswa.create-import', ['title' => $title]);
+    $tahun = TahunMasuk::latest()->get();
+    $data_category = Category::with('kelas')->get();
+    $data_kelas = Kelas::latest()->get();
+    return view('masterdata::siswa.create-import', ['tahun' => $tahun, 'title' => $title, 'data_category' => $data_category, 'data_kelas' => $data_kelas]);
   }
 
   public function import(Request $request)
   {
-    // Validasi file yang diupload
     $validator = Validator::make($request->all(), [
       'file' => 'required|mimes:xlsx,csv,xls',
+      'angkatan' => 'required',
+      'kelas_id' => 'required',
     ]);
 
-    // Jika validasi gagal, kembalikan dengan error
     if ($validator->fails()) {
       return redirect()->back()->withErrors($validator)->withInput();
     }
 
-    // Jika validasi sukses, lanjutkan dengan proses import
-    // Misalnya, menggunakan SiswaImport untuk mengimpor data
+    $kelas_id = $request->input('kelas_id');
+
+    $kelas = Kelas::with('category')->findOrFail($kelas_id);
+    $category_id = $kelas->category->id;
+
     try {
-      Excel::import(new SiswaImport, $request->file('file'));
+      Excel::import(new SiswaImport($request->angkatan, $category_id, $kelas_id), $request->file('file'));
+
       Alert::success('Success', 'Data berhasil diimpor');
     } catch (\Exception $e) {
-      Alert::error('Oops...', 'Terjadi kesalahan saat mengimport data, mohon perbaiki data');
-      return redirect()->route('siswa.index');
+      Alert::error('Oops...', 'Terjadi kesalahan: ' . $e->getMessage());
+      return redirect()->route('siswa.createImport');
     }
 
     return redirect()->route('siswa.index');
+  }
+
+
+
+
+
+  public function export()
+  {
+    return Excel::download(new SiswaExport, 'siswa.xlsx');
+  }
+
+  public function exportKategori($id)
+  {
+    $category = Category::findOrFail($id);
+    return Excel::download(new SiswaCategoryExport($id), 'siswa-' . $category->name . '.xlsx');
   }
 
 }
